requirements.txt
streamlit
numpy
opencv-python-headless
Pillow
tensorflow
streamlit-image-coordinates
